requirements.txt
pandas==2.2.0
numpy==1.24.0
scikit-learn==1.3.0
torch==2.2.0
transformers==4.36.0
requests==2.31.0
streamlit==1.29.0
plotly==5.17.0
python-dotenv==1.0.0
joblib==1.3.0
fhir.resources==6.2.0
typing-extensions==4.0.0
regex==2023.0.0
loguru==0.7.2
tqdm==4.66.1